==> scripts/Physiology/older unused analyses/5.segmented_regressions.Rmd <==
---
title: "Segmented Regressions - LFM vs. MPa"
author: "Indra Boving"
date: "2022-12-02"
output: html_document
---

# Setup
```{r}
knitr::opts_chunk$set(echo = TRUE)
#lots of extra here, but oh well___ 
library(ggplot2)
library(gapminder)
library(tidyverse)
library(lme4)
library(here)
library(segmented)
library(nlme)
library(lubridate)
library(plotrix)
filter = dplyr::filter
here = here::here
select = dplyr::select

# Function to plot mixed effects model segmented regressions:
source(here::here("scripts", "scripts_functions", "plot_segmented_MEM.R"))
```
# ------------------------
# 1. Field Data
```{r}
field_data_all_4 <- read_csv(here("processed-data", "field_data_2020_2021.csv"), show_col_types = FALSE) %>% 
 # select(-mpa, -water_potential) %>% 
  mutate(water_potential = midday, 
         mpa = water_potential,
         plant = udpipe::unique_identifier(unique_id)) %>% 
 # group_by(spp, plant) %>% 
  mutate(maxLFM = wettest_lfm)  %>% 
  mutate(strategy = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "Isohydric", 
    spp == "PIJE" ~ "Isohydric", 
     spp == "CADE" ~ "Isohydric", 
    spp == "CECO" ~ "Anisohydric", 
    spp == "ARPA" ~ "Anisohydric", 
    spp == "QUKE" ~ "Anisohydric"
  )) %>% 
   mutate(strategy_binary = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "1", 
    spp == "ARPA" ~ "2", 
    spp == "CADE" ~ "1", 
    spp == "CECO" ~ "2", 
    spp == "PIJE" ~ "1", 
    spp == "QUKE" ~ "2"
  )) %>% 
   mutate(Species = case_when(
     spp == "ABCO" ~ "A. concolor",
    spp == "PIJE" ~ "P. jeffreyi", 
    spp == "CADE" ~ "C. decurrens", 
    spp == "CECO" ~ "C. cordulatus", 
    spp == "ARPA" ~ "A. patula", 
    spp == "QUKE" ~ "Q. kelloggii"
    )) %>% 
  mutate(type = case_when(
    spp == "PIJE" ~ "tree", 
    spp == "QUKE" ~ "tree", 
    spp == "CECO" ~ "shrub", 
    spp == "ARPA" ~ "shrub", 
    spp == "ABCO" ~ "tree", 
    spp == "CADE" ~ "tree"
  )) %>% 
  mutate(F.Group = case_when(
    spp == "ARPA" ~ "Angiosperm", 
    spp == "ABCO" ~ "Gymnosperm",
    spp == "CADE" ~ "Gymnosperm",
    spp == "CECO" ~ "Angiosperm",
    spp == "PIJE" ~ "Gymnosperm",
    spp == "QUKE" ~ "Angiosperm"
  )) %>% 
  filter(lfm < 300, 
         lfm > 70, 
         age == "new"
         ) %>% 
  mutate(upr_lfm = lfm_tlp + (sd_lfm/2), 
         lwr_lfm = lfm_tlp - (sd_lfm/2))

unique(field_data_all_4$year)
```

```{r}
field_data_angio <- field_data_all_4 %>% 
  filter(strategy_binary == 2)
  
field_data_gymno <- field_data_all_4 %>% 
  filter(strategy_binary == 1)

abco <- field_data_all_4 %>% 
  filter(spp == "ABCO")

pije <- field_data_all_4 %>% 
  filter(spp == "PIJE")

cade <- field_data_all_4 %>% 
  filter(spp == "CADE")

quke  <- field_data_all_4 %>% 
  filter(spp == "QUKE")

arpa <- field_data_all_4 %>% 
  filter(spp == "ARPA")

ceco <- field_data_all_4 %>% 
  filter(spp == "CECO")
```


# 1.1. LFM
```{r}
lfm_psi <- c(rep(NaN, 4))
lfm_se <- c(rep(NaN, 4))
seg_psi_df <- tibble::data_frame(lfm_psi, lfm_se)


lfm_psi <- c(rep(NaN, 4))
lfm_se <- c(rep(NaN, 4))
lfm_p <- c(rep(NaN, 4))
seg_psi_norand_df <- tibble::data_frame(lfm_psi, lfm_se, lfm_p)

#With random
field_segment_df <- field_data_all_4 %>% 
  drop_na(mpa, lfm) %>% 
  mutate(id = unique_id)

o_mod<-lme(mpa~lfm + spp + doy, random=~1|id, data=field_segment_df)

os_mod<-segmented.default(o_mod, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod)

#put into the df:
seg_psi_df[1] <- summary(os_mod)$psi[2]
seg_psi_df[2] <- summary(os_mod)$psi[3]

#No random
u_mod <- lm(mpa~lfm + spp + doy, data = field_segment_df)
u_mod_b <- segmented(u_mod, ~lfm)
summary.segmented(u_mod_b)

u_mod_davies <- davies.test(u_mod_b)
u_mod_davies

#put into the df:
seg_psi_norand_df[1] <- summary(u_mod_b)$psi[2]
seg_psi_norand_df[2] <- summary(u_mod_b)$psi[3]
seg_psi_norand_df[3] <- u_mod_davies$p.value


#Set up SE grey areas: 

seg_psi_df <- seg_psi_df %>% 
  mutate(lwr = lfm_psi - (lfm_se/2), 
         upr = lfm_psi + (lfm_se/2)) %>% 
  distinct()

seg_psi_norand_df <- seg_psi_norand_df %>% 
  mutate(lwr = lfm_psi - (lfm_se/2), 
         upr = lfm_psi + (lfm_se/2))


ggplot(aes(y = mpa, 
           x = lfm, 
           color = spp), 
           data = field_data_all_4) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_df$lwr, 
                  xmax = seg_psi_df$upr), 
              fill = "grey80", 
              color = "grey80")  +
   geom_vline(xintercept  = seg_psi_norand_df$lfm_psi, color = "red") +
  geom_vline(xintercept  = seg_psi_df$lfm_psi, linetype = "dotted") +
  geom_vline(xintercept  = field_data_angio$mean_lfm_fall_spp, colour = "blue") +
  geom_vline(xintercept  = field_data_gymno$mean_lfm_fall_spp, colour = "purple") +
  labs(title = "Segments are all spp.")


# ggplot(aes(x = mpa, 
#            y = lfm, 
#            color = spp), 
#            data = field_segment_df) +
#   geom_point()+
#   geom_ribbon(aes(xmin = seg_psi_norand_df$lwr, 
#                   xmax = seg_psi_norand_df$upr), 
#               fill = "red", 
#               color = "red", 
#               alpha = .5) +
#   geom_ribbon(aes(xmin = seg_psi_df$lwr, 
#                   xmax = seg_psi_df$upr), 
#               fill = "pink", 
#               color = "pink", 
#               alpha = .5)   +
#   geom_vline(xintercept  = seg_psi_df$lfm_psi, color = "pink", size = 2)  +
#    geom_vline(xintercept  = seg_psi_norand_df$lfm_psi, color = "red", size = 2) +
#   geom_vline(xintercept  = seg_psi_norand_df$lfm_psi, linetype = "dotted") +
#   #geom_vline(xintercept  = field_data$tlp, colour = "blue") +
#  # geom_vline(xintercept  = field_data_gymno$tlp, colour = "purple") +
#   geom_vline(xintercept  = mean_tlp, colour = "orange", size = 2) +
#   geom_ribbon(aes(xmin = lwr_tlp, 
#                   xmax = upr_tlp), 
#               fill = "orange", 
#               color = "orange", 
#               alpha = .5)  +
#   labs(title = "Angiosperms", 
#        #caption = "pink = est. breakpoint with no random effects, red = with random effects, orange = tlp"
#        x = "", 
#        y = "LFM") +
#   theme(legend.position = "none") +
#   xlim(-5, 0)
```

```{r}
field_data_above <- field_data_all_4 %>% 
  filter(lfm > seg_psi_norand_df$lfm_psi)

mod_all <- lm(water_potential~lfm, data = field_data_above)

lfm_psi <- seg_psi_norand_df %>% 
  mutate(lfm = lfm_psi) %>% 
  as_data_frame() %>% 
  distinct()

mpa_seg <- predict(mod_all, lfm_psi)
mpa_seg

#below
field_data_below <- field_data_all_4 %>% 
  filter(lfm < seg_psi_norand_df$lfm_psi)

mod_all <- lm(water_potential~lfm, data = field_data_below)

lfm_psi <- seg_psi_norand_df %>% 
  mutate(lfm = lfm_psi) %>% 
  as_data_frame() %>% 
  distinct()

mpa_seg <- predict(mod_all, lfm_psi)
mpa_seg
```


## Angiosperms

```{r}
#Data
field_segment_angio_df <- field_data_all_4 %>% 
  filter(F.Group == "Angiosperm") %>% 
  drop_na(mpa, lfm) %>% 
  mutate(id = unique_id) 
 # filter(lfm > 80, 
       #  lfm < 150)

#Blank dfs
lfm_psi <- c(rep(NaN, 4))
lfm_se <- c(rep(NaN, 4))
seg_psi_angio_df <- tibble::data_frame(lfm_psi, lfm_se)


lfm_psi <- c(rep(NaN, 4))
lfm_se <- c(rep(NaN, 4))
lfm_p <- c(rep(NaN, 4))
seg_psi_norand_angio_df <- tibble::data_frame(lfm_psi, lfm_se, lfm_p)


#With random

o_mod_angio<-lme(mpa~lfm + spp + doy, random=~1|id, data=field_segment_angio_df)

os_mod_angio<-segmented.default(o_mod_angio, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod_angio)

#put into the df:
seg_psi_angio_df[1] <- summary(os_mod_angio)$psi[2]
seg_psi_angio_df[2] <- summary(os_mod_angio)$psi[3]

#No random
u_mod_angio <- lm(mpa~lfm + spp + doy, data = field_segment_angio_df)
u_mod_angio_b <- segmented(u_mod_angio, ~lfm, psi = 111)
summary.segmented(u_mod_angio_b)
u_mod_angio_b$psi

u_mod_angio_davies <- davies.test(u_mod_angio_b)
u_mod_angio_davies
#put into the df:
seg_psi_norand_angio_df[1] <- summary(u_mod_angio_b)$psi[2]
seg_psi_norand_angio_df[2] <- summary(u_mod_angio_b)$psi[3]
seg_psi_norand_angio_df[3] <- u_mod_angio_davies$p.value


#Set up SE grey areas: 

seg_psi_angio_df <- seg_psi_angio_df %>% 
  mutate(lwr = lfm_psi - (lfm_se/2), 
         upr = lfm_psi + (lfm_se/2)) %>% 
  distinct() %>% 
  mutate(F.Group = "Angiosperm")

seg_psi_norand_angio_df <- seg_psi_norand_angio_df %>% 
  mutate(lwr = lfm_psi - (lfm_se/2), 
         upr = lfm_psi + (lfm_se/2)) %>% 
  mutate(F.Group = "Angiosperm")



ggplot(aes(y = mpa, 
           x = lfm, 
           color = spp), 
           data = field_segment_angio_df) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_angio_df$lwr, 
                  xmax = seg_psi_angio_df$upr), 
              fill = "grey80", 
              color = "grey80")  +
   geom_vline(xintercept  = seg_psi_norand_angio_df$lfm_psi, color = "purple") +
  geom_vline(xintercept  = seg_psi_angio_df$lfm_psi, linetype = "dotted") +
  geom_vline(xintercept  = arpa$mean_lfm_fall_spp, colour = "red") +
    geom_vline(xintercept  = ceco$mean_lfm_fall_spp, colour = "orange") +
   # geom_vline(xintercept  = abco$mean_lfm_fall_spp, colour = "orange") +
    geom_vline(xintercept  = quke$mean_lfm_fall_spp, colour = "blue") +
  geom_vline(xintercept  = 145, colour = "pink") +
  # geom_ribbon(aes(xmin = arpa$lwr_lfm, 
  #                 xmax = arpa$upr_lfm),
  #             color = "red", alpha = .3, inherit.aes = FALSE) +
 # geom_ribbon(aes(xmin = field_data_angio$lwr_lfm, 
              #     xmax = field_data_angio$upr_lfm, 
              #     group = spp,
              #     fill = spp), 
              # fill = "grey80", 
              # color = "grey80")  +
  labs(title = "Angiosperms")
```

## Gymnosperms

```{r}
#Data
field_segment_gymno_df <- field_data_all_4 %>% 
  filter(F.Group == "Gymnosperm") %>% 
  drop_na(mpa, lfm) %>% 
  mutate(id = unique_id)



#Blank dfs
lfm_psi <- c(rep(NaN, 4))
lfm_se <- c(rep(NaN, 4))
seg_psi_gymno_df <- tibble::data_frame(lfm_psi, lfm_se)


lfm_psi <- c(rep(NaN, 4))
lfm_se <- c(rep(NaN, 4))
lfm_p <- c(rep(NaN, 4))
seg_psi_norand_gymno_df <- tibble::data_frame(lfm_psi, lfm_se, lfm_p)


#With random

o_mod_gymno<-lme(mpa~lfm + spp + doy, random=~1|id, data=field_segment_gymno_df)

os_mod_gymno<-segmented.default(o_mod_gymno, ~lfm, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod_gymno)

#put into the df:
seg_psi_gymno_df[1] <- summary(os_mod_gymno)$psi[2]
seg_psi_gymno_df[2] <- summary(os_mod_gymno)$psi[3]

#No random
u_mod_gymno <- lm(mpa~lfm + spp + doy, data = field_segment_gymno_df)
u_mod_gymno_b <- segmented(u_mod_gymno, ~lfm)
summary.segmented(u_mod_gymno_b)

u_mod_gymno_davies <- davies.test(u_mod_gymno_b)
u_mod_gymno_davies

#put into the df:
seg_psi_norand_gymno_df[1] <- summary(u_mod_gymno_b)$psi[2]
seg_psi_norand_gymno_df[2] <- summary(u_mod_gymno_b)$psi[3]
seg_psi_norand_gymno_df[3] <- u_mod_gymno_davies$p.value


#Set up SE grey areas: 

seg_psi_gymno_df <- seg_psi_gymno_df %>% 
  mutate(lwr = lfm_psi - (lfm_se/2), 
         upr = lfm_psi + (lfm_se/2)) %>% 
  distinct() %>% 
  mutate(F.Group = "Gymnosperm")

seg_psi_norand_gymno_df <- seg_psi_norand_gymno_df %>% 
  mutate(lwr = lfm_psi - (lfm_se/2), 
         upr = lfm_psi + (lfm_se/2)) %>% 
  mutate(F.Group = "Gymnosperm") %>% 
  distinct()


ggplot(aes(y = mpa, 
           x = lfm, 
           color = spp), 
           data = field_segment_gymno_df) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_norand_gymno_df$lwr, 
                  xmax = seg_psi_norand_gymno_df$upr), 
              fill = "grey80", 
              color = "grey80")  +
   geom_vline(xintercept  = seg_psi_norand_gymno_df$lfm_psi, color = "red") +
  geom_vline(xintercept  = seg_psi_norand_gymno_df$lfm_psi, linetype = "dotted") +
  geom_vline(xintercept  = field_segment_gymno_df$mean_lfm_fall_spp, 
             #aes(color = field_segment_gymno_df$spp)
            # color = c("blue", "purple", "pink")
             ) +
   geom_vline(xintercept  = abco$mean_lfm_fall_spp, colour = "red") +
    geom_vline(xintercept  = cade$mean_lfm_fall_spp, colour = "orange") +
   # geom_vline(xintercept  = abco$mean_lfm_fall_spp, colour = "orange") +
    geom_vline(xintercept  = pije$mean_lfm_fall_spp, colour = "blue") +
  geom_vline(xintercept  = 102, colour = "pink") +
  labs(title = "Gymnosperms")
```


Combine angio and gymno back together: 
```{r}
seg_psi_fgroups_df <- bind_rows(seg_psi_angio_df, seg_psi_gymno_df
                                ) %>% 
  distinct()

seg_psi_norand_fgroups_df <- bind_rows(seg_psi_norand_angio_df, seg_psi_norand_gymno_df) %>% 
  distinct()


#Visualize:

ggplot(aes(y = mpa, 
           x = lfm, 
           color = spp), 
           data = field_data_all_4) +
  geom_point()+
  #gymnos
  geom_ribbon(aes(xmin = seg_psi_norand_gymno_df$lwr, 
                  xmax = seg_psi_norand_gymno_df$upr), 
              fill = "grey80", 
              color = "grey80")  +
   geom_vline(xintercept  = seg_psi_norand_gymno_df$lfm_psi, color = "red") +
  geom_vline(xintercept  = seg_psi_norand_gymno_df$lfm_psi, linetype = "dotted") +
  #angios
  geom_vline(xintercept  = field_data_all_4$lfm_tlp, colour = "blue") +
  geom_ribbon(aes(xmin = seg_psi_angio_df$lwr, 
                  xmax = seg_psi_angio_df$upr), 
              fill = "grey", 
              color = "grey")  +
   geom_vline(xintercept  = seg_psi_norand_angio_df$lfm_psi, color = "pink") +
  geom_vline(xintercept  = seg_psi_norand_angio_df$lfm_psi, linetype = "dotted") +
  geom_vline(xintercept  = field_data_all_4$lfm_tlp, colour = "blue") +
  facet_wrap(~F.Group)
```

# 1.2. MPa

```{r}
mpa_psi <- c(rep(NaN, 4))
mpa_se <- c(rep(NaN, 4))
seg_psi_df <- tibble::data_frame(mpa_psi, mpa_se)


mpa_psi <- c(rep(NaN, 4))
mpa_se <- c(rep(NaN, 4))
mpa_p <- c(rep(NaN, 4))
seg_psi_norand_df <- tibble::data_frame(mpa_psi, mpa_se, mpa_p)

#With random
field_segment_df <- field_data_all_4 %>% 
  drop_na(mpa, lfm) %>% 
  mutate(id = unique_id)

o_mod<-lme(lfm ~ mpa + spp + doy, random=~1|id, data=field_segment_df)

os_mod<-segmented.default(o_mod, ~mpa, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod)

#put into the df:
seg_psi_df[1] <- summary(os_mod)$psi[2]
seg_psi_df[2] <- summary(os_mod)$psi[3]

#No random
u_mod <- lm(lfm~mpa+ spp + doy, data = field_segment_df)
u_mod_b <- segmented(u_mod, ~mpa)
summary.segmented(u_mod_b)

u_mod_davies <- davies.test(u_mod_b)
u_mod_davies

#put into the df:
seg_psi_norand_df[1] <- summary(u_mod_b)$psi[2]
seg_psi_norand_df[2] <- summary(u_mod_b)$psi[3]
seg_psi_norand_df[3] <- u_mod_davies$p.value


#Set up SE grey areas: 

seg_psi_df <- seg_psi_df %>% 
  mutate(lwr = mpa_psi - (mpa_se/2), 
         upr = mpa_psi + (mpa_se/2)) %>% 
  distinct()

seg_psi_norand_df <- seg_psi_norand_df %>% 
  mutate(lwr = mpa_psi - (mpa_se/2), 
         upr = mpa_psi + (mpa_se/2)) %>% 
  distinct()

mean_tlp <- field_data_all_4 %>% 
  drop_na(tlp) %>% 
 # filter(year == 2021, timing == "fall") %>% 
  mutate(mean_tlp_all = mean(tlp)) %>% 
  select(mean_tlp_all) %>% 
  distinct() %>% 
  pull()
mean_tlp

sd_tlp <- field_data_all_4 %>% 
  drop_na(tlp) %>% 
 # filter(year == 2021, timing == "fall") %>% 
  mutate(sd_tlp_all = sd(tlp)) %>% 
  select(sd_tlp_all) %>% 
  distinct() %>% 
  pull()
sd_tlp

upr_tlp <- mean(mean_tlp + (sd_tlp/2))
lwr_tlp <- mean(mean_tlp - (sd_tlp/2))

ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_data_all_4) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_norand_df$lwr, 
                  xmax = seg_psi_norand_df$upr), 
              fill = "red", 
              color = "red", 
              alpha = .5)  +
   geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, color = "red", size = 2) +
  geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, linetype = "dotted") +
  #geom_vline(xintercept  = field_data_angio$tlp, colour = "blue") +
 # geom_vline(xintercept  = field_data_gymno$tlp, colour = "purple") +
  geom_vline(xintercept  = mean_tlp, colour = "orange", size = 2) +
  geom_ribbon(aes(xmin = lwr_tlp, 
                  xmax = upr_tlp), 
              fill = "orange", 
              color = "orange", 
              alpha = .5)  +
  labs(title = "Segments are all spp.")
```


```{r}
field_data_all_4 %>%
   mutate(spp = fct_relevel(spp, "ABCO","PIJE","CADE","CECO", "ARPA","QUKE"))  %>% 
   mutate(Species = fct_relevel(Species,
                              "A. concolor",
                              "P. jeffreyi",
                              "C. decurrens",
                              "C. cordulatus",
                              "A. patula", 
                              "Q. kelloggii"
    )) %>% 
  filter(age != "both") %>% 
  mutate(month = month(date, label = TRUE, abbr = TRUE)) %>% 
  ggplot(aes(x = midday, 
             y = lfm, 
             color = as.factor(month), 
             shape = age)) +
  geom_point(size = 1, alpha = .7) +
  geom_vline(aes(xintercept = tlp), size = .5, alpha = .6,
             color = "black", linetype = "dotted") +
   geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, color = "red", size = 2) +
  #stat_smooth(aes(x = LFM, y = predawn, colour = Date, linetype = "Linear Fit"), method = "lm") 
  xlim(-5,0) + 
  labs( y = "Live Fuel Moisture", 
        x = "Midday (-MPa)", 
        color = "Month", 
        shape = "Age") +
  ylim(50,300) +
  facet_wrap(~strategy)
```

## Angiosperms
```{r}
#Data
field_segment_angio_df <- field_data_all_4 %>% 
  filter(F.Group == "Angiosperm") %>% 
  mutate(mpa = water_potential) %>% 
  #drop_na(mpa, lfm) %>% 
  mutate(id = unique_id) %>% 
  filter(mpa < 0)
 # filter(lfm > 80, 
       #  lfm < 150)

mpa_psi_angio <- c(rep(NaN, 4))
mpa_se_angio <- c(rep(NaN, 4))
seg_psi_df_angio <- tibble::data_frame(mpa_psi_angio, mpa_se_angio)


mpa_psi_angio <- c(rep(NaN, 4))
mpa_se_angio <- c(rep(NaN, 4))
mpa_p_angio <- c(rep(NaN, 4))
seg_psi_norand_df_angio <- tibble::data_frame(mpa_psi_angio, mpa_se_angio, mpa_p_angio)

#With random
field_segment_angio_nona_df <- field_segment_angio_df %>% 
  drop_na(lfm, mpa)

o_mod<-lme(lfm ~ mpa + spp + doy, random=~1|id, data=field_segment_angio_nona_df)

os_mod<-segmented.default(o_mod, ~mpa, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod)

#put into the df:
seg_psi_df_angio[1] <- summary(os_mod)$psi[2]
seg_psi_df_angio[2] <- summary(os_mod)$psi[3]

#No random
u_mod <- lm(lfm~mpa+ spp + doy, data = field_segment_angio_nona_df)
u_mod_b <- segmented(u_mod, ~mpa)
summary.segmented(u_mod_b)
summary(u_mod_b)

slopes_angio <- slope(u_mod_b)

u_mod_davies <- davies.test(u_mod, seg.Z = ~mpa)
u_mod_davies

u_mod_pscore <- pscore.test(u_mod, seg.Z=~mpa, k = 100, n.break = 1)
u_mod_pscore

#put into the df:
seg_psi_norand_df_angio[1] <- summary(u_mod_b)$psi[2]
seg_psi_norand_df_angio[2] <- summary(u_mod_b)$psi[3]
seg_psi_norand_df_angio[3] <- u_mod_davies$p.value


#Set up SE grey areas: 

seg_psi_df_angio <- seg_psi_df_angio %>% 
  mutate(lwr_angio = mpa_psi_angio - (mpa_se_angio/2), 
         upr_angio = mpa_psi_angio + (mpa_se_angio/2)) %>% 
  distinct()

seg_psi_norand_df_angio <- seg_psi_norand_df_angio %>% 
  mutate(lwr_angio = mpa_psi_angio - (mpa_se_angio/2), 
         upr_angio = mpa_psi_angio + (mpa_se_angio/2)) %>% 
  distinct()

mean_tlp_angio <- field_segment_angio_df %>% 
  drop_na(tlp) %>% 
 # filter(year == 2021, timing == "fall") %>% 
  mutate(mean_tlp_all = mean(tlp)) %>% 
  select(mean_tlp_all) %>% 
  distinct() %>% 
  pull()
mean_tlp_angio

sd_tlp_angio <- field_segment_angio_df %>% 
  drop_na(tlp) %>% 
 # filter(year == 2021, timing == "fall") %>% 
  mutate(sd_tlp_all = sd(tlp)) %>% 
  select(sd_tlp_all) %>% 
  distinct() %>% 
  pull() 
sd_tlp_angio

upr_tlp_angio <- mean(mean_tlp_angio + (sd_tlp_angio/2))
upr_tlp_angio
lwr_tlp_angio <- mean(mean_tlp_angio - (sd_tlp_angio/2))

angio_seg_plot <- ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_segment_angio_df) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_norand_df_angio$lwr_angio, 
                  xmax = seg_psi_norand_df_angio$upr_angio), 
              fill = "red", 
              color = "red", 
              alpha = .5) +
  geom_ribbon(aes(xmin = seg_psi_df_angio$lwr_angio, 
                  xmax = seg_psi_df_angio$upr_angio), 
              fill = "pink", 
              color = "pink", 
              alpha = .5)   +
  geom_vline(xintercept  = seg_psi_df_angio$mpa_psi_angio, color = "pink", size = 2)  +
   geom_vline(xintercept  = seg_psi_norand_df_angio$mpa_psi_angio, color = "red", size = 2) +
  geom_vline(xintercept  = seg_psi_norand_df_angio$mpa_psi_angio, linetype = "dotted") +
  #geom_vline(xintercept  = field_data_angio$tlp, colour = "blue") +
 # geom_vline(xintercept  = field_data_gymno$tlp, colour = "purple") +
  geom_vline(xintercept  = mean_tlp_angio, colour = "orange", size = 2) +
  geom_ribbon(aes(xmin = lwr_tlp_angio, 
                  xmax = upr_tlp_angio), 
              fill = "orange", 
              color = "orange", 
              alpha = .5)  +
  labs(title = "Angiosperms", 
       #caption = "pink = est. breakpoint with no random effects, red = with random effects, orange = tlp"
       x = "", 
       y = "LFM") +
  theme(legend.position = "none") +
  xlim(-5, 0)

angio_seg_plot
```

## Gymnosperms
```{r}
#Data
field_segment_gymno_df <- field_data_all_4 %>% 
  filter(F.Group == "Gymnosperm") %>% 
  mutate(id = unique_id,
         mpa = water_potential) %>% 
  drop_na(mpa, lfm)

mpa_psi <- c(rep(NaN, 4))
mpa_se <- c(rep(NaN, 4))
seg_psi_df <- tibble::data_frame(mpa_psi, mpa_se)


mpa_psi <- c(rep(NaN, 4))
mpa_se <- c(rep(NaN, 4))
mpa_p <- c(rep(NaN, 4))
seg_psi_norand_df <- tibble::data_frame(mpa_psi, mpa_se, mpa_p)

#With random
field_segment_gymno_df <- field_segment_gymno_df %>% 
  drop_na(mpa, lfm) %>% 
  mutate(id = unique_id)

o_mod<-lme(lfm ~ mpa + spp + doy, random=~1|id, data=field_segment_gymno_df)

os_mod<-segmented.default(o_mod, ~mpa, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod)

#put into the df:
seg_psi_df[1] <- summary(os_mod)$psi[2]
seg_psi_df[2] <- summary(os_mod)$psi[3]

#No random
u_mod <- lm(lfm~mpa+ spp + doy, data = field_segment_gymno_df)
u_mod_b <- segmented(u_mod, ~mpa)
summary.segmented(u_mod_b)

slopes_gymno <- slope(u_mod_b)

u_mod_davies <- davies.test(u_mod, seg.Z = ~mpa)
u_mod_davies

u_mod_pscore <- pscore.test(u_mod_b)
u_mod_pscore

#put into the df:
seg_psi_norand_df[1] <- summary(u_mod_b)$psi[2]
seg_psi_norand_df[2] <- summary(u_mod_b)$psi[3]
seg_psi_norand_df[3] <- u_mod_davies$p.value


#Set up SE grey areas: 

seg_psi_df <- seg_psi_df %>% 
  mutate(lwr = mpa_psi - (mpa_se/2), 
         upr = mpa_psi + (mpa_se/2)) %>% 
  distinct()

seg_psi_norand_df <- seg_psi_norand_df %>% 
  mutate(lwr = mpa_psi - (mpa_se/2), 
         upr = mpa_psi + (mpa_se/2)) %>% 
  distinct()

mean_tlp <- field_segment_gymno_df %>% 
  drop_na(tlp) %>% 
 # filter(year == 2021, timing == "fall") %>% 
  mutate(mean_tlp_all = mean(tlp)) %>% 
  select(mean_tlp_all) %>% 
  distinct() %>% 
  pull()
mean_tlp

sd_tlp <- field_segment_gymno_df %>% 
  drop_na(tlp) %>% 
 # filter(year == 2021, timing == "fall") %>% 
  mutate(sd_tlp_all = sd(tlp)) %>% 
  select(sd_tlp_all) %>% 
  distinct() %>% 
  pull() 
sd_tlp

upr_tlp <- mean(mean_tlp + (sd_tlp/2))
upr_tlp
lwr_tlp <- mean(mean_tlp - (sd_tlp/2))

gymno_seg_plot <- ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_segment_gymno_df) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_norand_df$lwr, 
                  xmax = seg_psi_norand_df$upr), 
              fill = "red", 
              color = "red", 
              alpha = .5) +
 geom_ribbon(aes(xmin = seg_psi_df$lwr,
                 xmax = seg_psi_df$upr),
             fill = "pink",
             color = "pink",
             alpha = .5)   +
 geom_vline(xintercept  = seg_psi_df$mpa_psi, color = "pink", size = 2) +
   geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, color = "red", size = 2) +
  geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, linetype = "dotted") +
  #geom_vline(xintercept  = field_data_gymno$tlp, colour = "blue") +
 # geom_vline(xintercept  = field_data_gymno$tlp, colour = "purple") +
  geom_vline(xintercept  = mean_tlp, colour = "orange", size = 2) +
  geom_ribbon(aes(xmin = lwr_tlp, 
                  xmax = upr_tlp), 
              fill = "orange", 
              color = "orange", 
              alpha = .5)  +
  labs(title = "Gymnosperms", 
       caption = "pink = est. breakpoint with no random effects, red = with random effects, orange = tlp") +
  theme(legend.position = "none") +
  xlim(-5, 0)

gymno_seg_plot
```

Figure back together: 
```{r}
legend_gymno <- cowplot::get_legend(ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_segment_df) +
  geom_point()+
   geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, color = "red", size = 2) +
  geom_vline(xintercept  = mean_tlp, colour = "orange", size = 2) +
  labs(color = "Species"))

plot_1 <- cowplot::plot_grid(angio_seg_plot, gymno_seg_plot, ncol = 1)
plot_1

plot_legend <- cowplot::plot_grid(plot_1, legend, ncol = 2, rel_widths = c(4,1))
plot_legend
```

These are the slopes for before and after the breakpoints:
```{r}
slopes_angio
slopes_gymno
```

# 1.3. -1/MPa

## Angiosperms 
```{r}
#Data
field_segment_angio_df <- field_data_all_4 %>% 
  filter(F.Group == "Angiosperm") %>% 
  mutate(mpa = -1/water_potential) %>% 
  #drop_na(mpa, lfm) %>% 
  mutate(id = unique_id) 
 # filter(lfm > 80, 
       #  lfm < 150)

mpa_psi_angio <- c(rep(NaN, 4))
mpa_se_angio <- c(rep(NaN, 4))
seg_psi_df_angio <- tibble::data_frame(mpa_psi_angio, mpa_se_angio)


mpa_psi_angio <- c(rep(NaN, 4))
mpa_se_angio <- c(rep(NaN, 4))
mpa_p_angio <- c(rep(NaN, 4))
seg_psi_norand_df_angio <- tibble::data_frame(mpa_psi_angio, mpa_se_angio, mpa_p_angio)

#With random
field_segment_angio_nona_df <- field_segment_angio_df %>% 
  drop_na(lfm, mpa)

o_mod<-lme(lfm ~ mpa + spp + doy, random=~1|id, data=field_segment_angio_nona_df)

os_mod<-segmented.default(o_mod, ~mpa, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod)

#put into the df:
seg_psi_df_angio[1] <- summary(os_mod)$psi[2]
seg_psi_df_angio[2] <- summary(os_mod)$psi[3]

#No random
u_mod <- lm(lfm~mpa+ spp + doy, data = field_segment_angio_nona_df)
u_mod_b <- segmented(u_mod, ~mpa)
summary.segmented(u_mod_b)

u_mod_davies <- davies.test(u_mod_b)
u_mod_davies

#put into the df:
seg_psi_norand_df_angio[1] <- summary(u_mod_b)$psi[2]
seg_psi_norand_df_angio[2] <- summary(u_mod_b)$psi[3]
seg_psi_norand_df_angio[3] <- u_mod_davies$p.value


#Set up SE grey areas: 

seg_psi_df_angio <- seg_psi_df_angio %>% 
  mutate(lwr_angio = mpa_psi_angio - (mpa_se_angio/2), 
         upr_angio = mpa_psi_angio + (mpa_se_angio/2)) %>% 
  distinct()

seg_psi_norand_df_angio <- seg_psi_norand_df_angio %>% 
  mutate(lwr_angio = mpa_psi_angio - (mpa_se_angio/2), 
         upr_angio = mpa_psi_angio + (mpa_se_angio/2)) %>% 
  distinct()

mean_tlp_angio <- field_segment_angio_df %>% 
  drop_na(tlp) %>% 
 # filter(year == 2021, timing == "fall") %>% 
  mutate(mean_tlp_all = mean(tlp)) %>% 
  select(mean_tlp_all) %>% 
  distinct() %>% 
  pull()
mean_tlp_angio

sd_tlp_angio <- field_segment_angio_df %>% 
  drop_na(tlp) %>% 
 # filter(year == 2021, timing == "fall") %>% 
  mutate(sd_tlp_all = sd(tlp)) %>% 
  select(sd_tlp_all) %>% 
  distinct() %>% 
  pull() 
sd_tlp_angio

upr_tlp_angio <- mean(mean_tlp_angio + (sd_tlp_angio/2))
upr_tlp_angio
lwr_tlp_angio <- mean(mean_tlp_angio - (sd_tlp_angio/2))

angio_seg_plot <- ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_segment_angio_df) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_norand_df_angio$lwr_angio, 
                  xmax = seg_psi_norand_df_angio$upr_angio), 
              fill = "red", 
              color = "red", 
              alpha = .5) +
  geom_ribbon(aes(xmin = seg_psi_df_angio$lwr_angio, 
                  xmax = seg_psi_df_angio$upr_angio), 
              fill = "pink", 
              color = "pink", 
              alpha = .5)   +
  geom_vline(xintercept  = seg_psi_df_angio$mpa_psi_angio, color = "pink", size = 2)  +
   geom_vline(xintercept  = seg_psi_norand_df_angio$mpa_psi_angio, color = "red", size = 2) +
  geom_vline(xintercept  = seg_psi_norand_df_angio$mpa_psi_angio, linetype = "dotted") +
  #geom_vline(xintercept  = field_data_angio$tlp, colour = "blue") +
 # geom_vline(xintercept  = field_data_gymno$tlp, colour = "purple") +
  geom_vline(xintercept  = -1/mean_tlp_angio, colour = "orange", size = 2) +
  geom_ribbon(aes(xmin = -1/lwr_tlp_angio, 
                  xmax = -1/upr_tlp_angio), 
              fill = "orange", 
              color = "orange", 
              alpha = .5)  +
  labs(title = "Angiosperms", 
       #caption = "pink = est. breakpoint with no random effects, red = with random effects, orange = tlp"
       x = "", 
       y = "LFM") +
  theme(legend.position = "none") 

angio_seg_plot
```

## Gymnosperms
```{r}
#Data
field_segment_gymno_df <- field_data_all_4 %>% 
  filter(F.Group == "Gymnosperm") %>% 
  mutate(id = unique_id,
         mpa = -1/water_potential) %>% 
  drop_na(mpa, lfm)

mpa_psi <- c(rep(NaN, 4))
mpa_se <- c(rep(NaN, 4))
seg_psi_df <- tibble::data_frame(mpa_psi, mpa_se)


mpa_psi <- c(rep(NaN, 4))
mpa_se <- c(rep(NaN, 4))
mpa_p <- c(rep(NaN, 4))
seg_psi_norand_df <- tibble::data_frame(mpa_psi, mpa_se, mpa_p)

#With random
field_segment_gymno_df <- field_segment_gymno_df %>% 
  drop_na(mpa, lfm) %>% 
  mutate(id = unique_id)

o_mod<-lme(lfm ~ mpa + spp + doy, random=~1|id, data=field_segment_gymno_df)

os_mod<-segmented.default(o_mod, ~mpa, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod)

#put into the df:
seg_psi_df[1] <- summary(os_mod)$psi[2]
seg_psi_df[2] <- summary(os_mod)$psi[3]

#No random
u_mod <- lm(lfm~mpa+ spp + doy, data = field_segment_gymno_df)
u_mod_b <- segmented(u_mod, ~mpa)
summary.segmented(u_mod_b)

u_mod_davies <- davies.test(u_mod_b)
u_mod_davies

#put into the df:
seg_psi_norand_df[1] <- summary(u_mod_b)$psi[2]
seg_psi_norand_df[2] <- summary(u_mod_b)$psi[3]
seg_psi_norand_df[3] <- u_mod_davies$p.value


#Set up SE grey areas: 

seg_psi_df <- seg_psi_df %>% 
  mutate(lwr = mpa_psi - (mpa_se/2), 
         upr = mpa_psi + (mpa_se/2)) %>% 
  distinct()

seg_psi_norand_df <- seg_psi_norand_df %>% 
  mutate(lwr = mpa_psi - (mpa_se/2), 
         upr = mpa_psi + (mpa_se/2)) %>% 
  distinct()

mean_tlp <- field_segment_gymno_df %>% 
  drop_na(tlp) %>% 
 # filter(year == 2021, timing == "fall") %>% 
  mutate(mean_tlp_all = mean(tlp)) %>% 
  select(mean_tlp_all) %>% 
  distinct() %>% 
  pull()
mean_tlp

sd_tlp <- field_segment_gymno_df %>% 
  drop_na(tlp) %>% 
 # filter(year == 2021, timing == "fall") %>% 
  mutate(sd_tlp_all = sd(tlp)) %>% 
  select(sd_tlp_all) %>% 
  distinct() %>% 
  pull() 
sd_tlp

upr_tlp <- mean(mean_tlp + (sd_tlp/2))
upr_tlp
lwr_tlp <- mean(mean_tlp - (sd_tlp/2))

gymno_seg_plot <- ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_segment_gymno_df) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_norand_df$lwr, 
                  xmax = seg_psi_norand_df$upr), 
              fill = "red", 
              color = "red", 
              alpha = .5) +
 geom_ribbon(aes(xmin = seg_psi_df$lwr,
                 xmax = seg_psi_df$upr),
             fill = "pink",
             color = "pink",
             alpha = .5)   +
 geom_vline(xintercept  = seg_psi_df$mpa_psi, color = "pink", size = 2) +
   geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, color = "red", size = 2) +
  geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, linetype = "dotted") +
  #geom_vline(xintercept  = field_data_gymno$tlp, colour = "blue") +
 # geom_vline(xintercept  = field_data_gymno$tlp, colour = "purple") +
  geom_vline(xintercept  = -1/mean_tlp, colour = "orange", size = 2) +
  geom_ribbon(aes(xmin = -1/lwr_tlp, 
                  xmax = -1/upr_tlp), 
              fill = "orange", 
              color = "orange", 
              alpha = .5)  +
  labs(title = "Gymnosperms", 
       caption = "pink = est. breakpoint with no random effects, red = with random effects, orange = tlp") +
  theme(legend.position = "none") +
  xlim(0, 1)

gymno_seg_plot

```
Figure back together: 
```{r}
legend_gymno <- cowplot::get_legend(ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_segment_df) +
  geom_point()+
   geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, color = "red", size = 2) +
  geom_vline(xintercept  = mean_tlp, colour = "orange", size = 2) +
  labs(color = "Species"))

plot_1 <- cowplot::plot_grid(angio_seg_plot, gymno_seg_plot, ncol = 1)
plot_1

plot_legend <- cowplot::plot_grid(plot_1, legend, ncol = 2, rel_widths = c(4,1))
plot_legend
```

#------------------------
# 2. PV Curves
```{r,warning=FALSE}
pv_all_df_4 <- read_csv(here("processed-data", "all_pv_curves_clean.csv"), show_col_types = FALSE) %>%
  mutate(spp == species) %>% 
  filter(!spp %in% c("ADFA", "CEME")) %>% 
  mutate(sample = as.factor(sample)) %>% 
  mutate(month = as.factor(month)) %>% 
  mutate(date = ymd(date), 
        individual = str_c(species, '_', sample, '_', year)
        ) %>% 
  mutate(year_month = str_c(year, '_', month)) %>% 
  clean_names() %>% 
  group_by(spp) %>% 
  mutate(plant = udpipe::unique_identifier(individual), 
         fresh_weight_saturated = as.numeric(fresh_weight_saturated)) %>% 
  group_by(spp, plant) %>% 
  mutate(maxLFM = ((fresh_weight_saturated - dry_weight)/dry_weight)*100) %>% 
   mutate(spp = fct_relevel(spp, "ABCO","PIJE","CADE","CECO", "ARPA","QUKE")) %>% 
  mutate(strategy = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "Isohydric", 
    spp == "PIJE" ~ "Isohydric", 
     spp == "CADE" ~ "Isohydric", 
    spp == "CECO" ~ "Anisohydric", 
    spp == "ARPA" ~ "Anisohydric", 
    spp == "QUKE" ~ "Anisohydric"
  )) %>% 
   mutate(strategy_binary = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "1", 
    spp == "ARPA" ~ "2", 
    spp == "CADE" ~ "1", 
    spp == "CECO" ~ "2", 
    spp == "PIJE" ~ "1", 
    spp == "QUKE" ~ "2"
  )) %>% 
   mutate(Species = case_when(
     spp == "ABCO" ~ "A. concolor",
    spp == "PIJE" ~ "P. jeffreyi", 
    spp == "CADE" ~ "C. decurrens", 
    spp == "CECO" ~ "C. cordulatus", 
    spp == "ARPA" ~ "A. patula", 
    spp == "QUKE" ~ "Q. kelloggii"
    )) %>% 
  mutate(type = case_when(
    spp == "PIJE" ~ "tree", 
    spp == "QUKE" ~ "tree", 
    spp == "CECO" ~ "shrub", 
    spp == "ARPA" ~ "shrub", 
    spp == "ABCO" ~ "tree", 
    spp == "CADE" ~ "tree"
  )) %>% 
  mutate(F.Group = case_when(
    spp == "ARPA" ~ "Angiosperm", 
    spp == "ABCO" ~ "Gymnosperm",
    spp == "CADE" ~ "Gymnosperm",
    spp == "CECO" ~ "Angiosperm",
    spp == "PIJE" ~ "Gymnosperm",
    spp == "QUKE" ~ "Angiosperm"
  )) %>% 
  filter(timing == "fall")
  
```

## Angiosperms 
```{r}
#Data
field_segment_angio_df <- pv_all_df_4 %>% 
  filter(F.Group == "Angiosperm") %>% 
  mutate(mpa = water_potential) %>% 
  #drop_na(mpa, lfm) %>% 
  mutate(id = plant) %>% 
  filter(mpa < 0)
 # filter(lfm > 80, 
       #  lfm < 150)

mpa_psi_angio <- c(rep(NaN, 4))
mpa_se_angio <- c(rep(NaN, 4))
seg_psi_df_angio <- tibble::data_frame(mpa_psi_angio, mpa_se_angio)


mpa_psi_angio <- c(rep(NaN, 4))
mpa_se_angio <- c(rep(NaN, 4))
mpa_p_angio <- c(rep(NaN, 4))
seg_psi_norand_df_angio <- tibble::data_frame(mpa_psi_angio, mpa_se_angio, mpa_p_angio)

#With random
field_segment_angio_nona_df <- field_segment_angio_df %>% 
  drop_na(lfm, mpa)

o_mod<-lme(lfm ~ mpa + spp, random=~1|id, data=field_segment_angio_nona_df)

os_mod<-segmented.default(o_mod, ~mpa, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod)

#put into the df:
seg_psi_df_angio[1] <- summary(os_mod)$psi[2]
seg_psi_df_angio[2] <- summary(os_mod)$psi[3]

#No random
u_mod <- lm(lfm~mpa+ spp, data = field_segment_angio_nona_df)
u_mod_b <- segmented(u_mod, ~mpa)
summary.segmented(u_mod_b)
summary(u_mod_b)

slopes_angio <- slope(u_mod_b)

u_mod_davies <- davies.test(u_mod, seg.Z = ~mpa)
u_mod_davies

u_mod_pscore <- pscore.test(u_mod, seg.Z=~mpa, k = 100, n.break = 1)
u_mod_pscore

#put into the df:
seg_psi_norand_df_angio[1] <- summary(u_mod_b)$psi[2]
seg_psi_norand_df_angio[2] <- summary(u_mod_b)$psi[3]
seg_psi_norand_df_angio[3] <- u_mod_davies$p.value


#Set up SE grey areas: 

seg_psi_df_angio <- seg_psi_df_angio %>% 
  mutate(lwr_angio = mpa_psi_angio - (mpa_se_angio/2), 
         upr_angio = mpa_psi_angio + (mpa_se_angio/2)) %>% 
  distinct()

seg_psi_norand_df_angio <- seg_psi_norand_df_angio %>% 
  mutate(lwr_angio = mpa_psi_angio - (mpa_se_angio/2), 
         upr_angio = mpa_psi_angio + (mpa_se_angio/2)) %>% 
  distinct()

# mean_tlp_angio <- field_segment_angio_df %>% 
#   drop_na(tlp) %>% 
#  # filter(year == 2021, timing == "fall") %>% 
#   mutate(mean_tlp_all = mean(tlp)) %>% 
#   select(mean_tlp_all) %>% 
#   distinct() %>% 
#   pull()
# mean_tlp_angio

# sd_tlp_angio <- field_segment_angio_df %>% 
#   drop_na(tlp) %>% 
#  # filter(year == 2021, timing == "fall") %>% 
#   mutate(sd_tlp_all = sd(tlp)) %>% 
#   select(sd_tlp_all) %>% 
#   distinct() %>% 
#   pull() 
# sd_tlp_angio

# upr_tlp_angio <- mean(mean_tlp_angio + (sd_tlp_angio/2))
# upr_tlp_angio
# lwr_tlp_angio <- mean(mean_tlp_angio - (sd_tlp_angio/2))

angio_seg_plot <- ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_segment_angio_df) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_norand_df_angio$lwr_angio, 
                  xmax = seg_psi_norand_df_angio$upr_angio), 
              fill = "red", 
              color = "red", 
              alpha = .5) +
  geom_ribbon(aes(xmin = seg_psi_df_angio$lwr_angio, 
                  xmax = seg_psi_df_angio$upr_angio), 
              fill = "pink", 
              color = "pink", 
              alpha = .5)   +
  geom_vline(xintercept  = seg_psi_df_angio$mpa_psi_angio, color = "pink", size = 2)  +
   geom_vline(xintercept  = seg_psi_norand_df_angio$mpa_psi_angio, color = "red", size = 2) +
  geom_vline(xintercept  = seg_psi_norand_df_angio$mpa_psi_angio, linetype = "dotted") +
  #geom_vline(xintercept  = field_data_angio$tlp, colour = "blue") +
 # geom_vline(xintercept  = field_data_gymno$tlp, colour = "purple") +
  geom_vline(xintercept  = mean_tlp_angio, colour = "orange", size = 2) +
  geom_ribbon(aes(xmin = lwr_tlp_angio, 
                  xmax = upr_tlp_angio), 
              fill = "orange", 
              color = "orange", 
              alpha = .5)  +
  labs(title = "Angiosperms", 
       #caption = "pink = est. breakpoint with no random effects, red = with random effects, orange = tlp"
       x = "", 
       y = "LFM") +
  theme(legend.position = "none") +
  xlim(-5, 0)

angio_seg_plot
```

## Gymnosperms
```{r}
#Data
field_segment_gymno_df <- pv_all_df_4 %>% 
  filter(F.Group == "Gymnosperm") %>% 
  mutate(id = plant,
         mpa = water_potential) %>% 
  drop_na(mpa, lfm)

mpa_psi <- c(rep(NaN, 4))
mpa_se <- c(rep(NaN, 4))
seg_psi_df <- tibble::data_frame(mpa_psi, mpa_se)


mpa_psi <- c(rep(NaN, 4))
mpa_se <- c(rep(NaN, 4))
mpa_p <- c(rep(NaN, 4))
seg_psi_norand_df <- tibble::data_frame(mpa_psi, mpa_se, mpa_p)

#With random
field_segment_gymno_df <- field_segment_gymno_df %>% 
  drop_na(mpa, lfm) %>% 
  mutate(id = plant)

o_mod<-lme(lfm ~ mpa + spp, random=~1|id, data=field_segment_gymno_df)

os_mod<-segmented.default(o_mod, ~mpa, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod)

#put into the df:
seg_psi_df[1] <- summary(os_mod)$psi[2]
seg_psi_df[2] <- summary(os_mod)$psi[3]

#No random
u_mod <- lm(lfm~mpa+ spp, data = field_segment_gymno_df)
u_mod_b <- segmented(u_mod, ~mpa)
summary.segmented(u_mod_b)

slopes_gymno <- slope(u_mod_b)

u_mod_davies <- davies.test(u_mod, seg.Z = ~mpa)
u_mod_davies

u_mod_pscore <- pscore.test(u_mod_b)
u_mod_pscore

#put into the df:
seg_psi_norand_df[1] <- summary(u_mod_b)$psi[2]
seg_psi_norand_df[2] <- summary(u_mod_b)$psi[3]
seg_psi_norand_df[3] <- u_mod_davies$p.value


#Set up SE grey areas: 

seg_psi_df <- seg_psi_df %>% 
  mutate(lwr = mpa_psi - (mpa_se/2), 
         upr = mpa_psi + (mpa_se/2)) %>% 
  distinct()

seg_psi_norand_df <- seg_psi_norand_df %>% 
  mutate(lwr = mpa_psi - (mpa_se/2), 
         upr = mpa_psi + (mpa_se/2)) %>% 
  distinct()

# mean_tlp <- field_segment_gymno_df %>% 
#   drop_na(tlp) %>% 
#  # filter(year == 2021, timing == "fall") %>% 
#   mutate(mean_tlp_all = mean(tlp)) %>% 
#   select(mean_tlp_all) %>% 
#   distinct() %>% 
#   pull()
# mean_tlp
# 
# sd_tlp <- field_segment_gymno_df %>% 
#   drop_na(tlp) %>% 
#  # filter(year == 2021, timing == "fall") %>% 
#   mutate(sd_tlp_all = sd(tlp)) %>% 
#   select(sd_tlp_all) %>% 
#   distinct() %>% 
#   pull() 
# sd_tlp
# 
# upr_tlp <- mean(mean_tlp + (sd_tlp/2))
# upr_tlp
# lwr_tlp <- mean(mean_tlp - (sd_tlp/2))

gymno_seg_plot <- ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_segment_gymno_df) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_norand_df$lwr, 
                  xmax = seg_psi_norand_df$upr), 
              fill = "red", 
              color = "red", 
              alpha = .5) +
 geom_ribbon(aes(xmin = seg_psi_df$lwr,
                 xmax = seg_psi_df$upr),
             fill = "pink",
             color = "pink",
             alpha = .5)   +
 geom_vline(xintercept  = seg_psi_df$mpa_psi, color = "pink", size = 2) +
   geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, color = "red", size = 2) +
  geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, linetype = "dotted") +
  #geom_vline(xintercept  = field_data_gymno$tlp, colour = "blue") +
 # geom_vline(xintercept  = field_data_gymno$tlp, colour = "purple") +
  geom_vline(xintercept  = mean_tlp, colour = "orange", size = 2) +
  geom_ribbon(aes(xmin = lwr_tlp, 
                  xmax = upr_tlp), 
              fill = "orange", 
              color = "orange", 
              alpha = .5)  +
  labs(title = "Gymnosperms", 
       caption = "pink = est. breakpoint with no random effects, red = with random effects, orange = tlp") +
  theme(legend.position = "none") +
  xlim(-5, 0)

gymno_seg_plot

```

# -------------------------

# 3. Flammability Drydown

```{r, warning=FALSE}
flam_curve_phys_all_raw <- read_csv(here("processed-data", "sierra_flam_data_all.csv"), show_col_types = FALSE)

flam_curve_df_4 <- flam_curve_phys_all_raw %>% 
   filter(!spp %in% c("ADFA", "CEME")) %>% 
  mutate(water_potential = -1*pos_mpa, 
         swc = swc_sat, 
         month = as.factor(month)) %>% 
  group_by(spp) %>% 
  mutate(plant = udpipe::unique_identifier(individual), 
         maxLFM = gww_gdw_saturated * 100) %>% 
  mutate(spp = toupper(spp)) %>% 
  mutate(strategy = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "Isohydric", 
    spp == "PIJE" ~ "Isohydric", 
     spp == "CADE" ~ "Isohydric", 
    spp == "CECO" ~ "Anisohydric", 
    spp == "ARPA" ~ "Anisohydric", 
    spp == "QUKE" ~ "Anisohydric"
  )) %>% 
   mutate(strategy_binary = case_when( #based on slope of predawn~midday line
    spp == "ABCO" ~ "1", 
    spp == "ARPA" ~ "2", 
    spp == "CADE" ~ "1", 
    spp == "CECO" ~ "2", 
    spp == "PIJE" ~ "1", 
    spp == "QUKE" ~ "2"
  )) %>% 
   mutate(Species = case_when(
     spp == "ABCO" ~ "A. concolor",
    spp == "PIJE" ~ "P. jeffreyi", 
    spp == "CADE" ~ "C. decurrens", 
    spp == "CECO" ~ "C. cordulatus", 
    spp == "ARPA" ~ "A. patula", 
    spp == "QUKE" ~ "Q. kelloggii"
    )) %>% 
  mutate(type = case_when(
    spp == "PIJE" ~ "tree", 
    spp == "QUKE" ~ "tree", 
    spp == "CECO" ~ "shrub", 
    spp == "ARPA" ~ "shrub", 
    spp == "ABCO" ~ "tree", 
    spp == "CADE" ~ "tree"
  )) %>% 
  mutate(F.Group = case_when(
    spp == "ARPA" ~ "Angiosperm", 
    spp == "ABCO" ~ "Gymnosperm",
    spp == "CADE" ~ "Gymnosperm",
    spp == "CECO" ~ "Angiosperm",
    spp == "PIJE" ~ "Gymnosperm",
    spp == "QUKE" ~ "Angiosperm"
  )) %>% 
  select(lfm, water_potential, mpa, year, month, year_month, plant, F.Group, spp)
#%>% 
 # select(spp, swc, lfm, mpa, month, tlp, water_potential, po, individual, plant)
```

## Angiosperms
```{r}
#Data
field_segment_angio_df <- flam_curve_df_4 %>% 
  filter(F.Group == "Angiosperm") %>% 
  mutate(mpa = water_potential) %>% 
  #drop_na(mpa, lfm) %>% 
  mutate(id = plant) %>% 
  filter(mpa < 0)
 # filter(lfm > 80, 
       #  lfm < 150)

mpa_psi_angio <- c(rep(NaN, 4))
mpa_se_angio <- c(rep(NaN, 4))
seg_psi_df_angio <- tibble::data_frame(mpa_psi_angio, mpa_se_angio)


mpa_psi_angio <- c(rep(NaN, 4))
mpa_se_angio <- c(rep(NaN, 4))
mpa_p_angio <- c(rep(NaN, 4))
seg_psi_norand_df_angio <- tibble::data_frame(mpa_psi_angio, mpa_se_angio, mpa_p_angio)

#With random
field_segment_angio_nona_df <- field_segment_angio_df %>% 
  drop_na(lfm, mpa)

o_mod<-lme(lfm ~ mpa + spp, random=~1|id, data=field_segment_angio_nona_df)

os_mod<-segmented.default(o_mod, ~mpa, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod)

#put into the df:
seg_psi_df_angio[1] <- summary(os_mod)$psi[2]
seg_psi_df_angio[2] <- summary(os_mod)$psi[3]

#No random
u_mod <- lm(lfm~mpa+ spp, data = field_segment_angio_nona_df)
u_mod_b <- segmented(u_mod, ~mpa)
summary.segmented(u_mod_b)
summary(u_mod_b)

slopes_angio <- slope(u_mod_b)

u_mod_davies <- davies.test(u_mod, seg.Z = ~mpa)
u_mod_davies

u_mod_pscore <- pscore.test(u_mod, seg.Z=~mpa, k = 100, n.break = 1)
u_mod_pscore

#put into the df:
seg_psi_norand_df_angio[1] <- summary(u_mod_b)$psi[2]
seg_psi_norand_df_angio[2] <- summary(u_mod_b)$psi[3]
seg_psi_norand_df_angio[3] <- u_mod_davies$p.value


#Set up SE grey areas: 

seg_psi_df_angio <- seg_psi_df_angio %>% 
  mutate(lwr_angio = mpa_psi_angio - (mpa_se_angio/2), 
         upr_angio = mpa_psi_angio + (mpa_se_angio/2)) %>% 
  distinct()

seg_psi_norand_df_angio <- seg_psi_norand_df_angio %>% 
  mutate(lwr_angio = mpa_psi_angio - (mpa_se_angio/2), 
         upr_angio = mpa_psi_angio + (mpa_se_angio/2)) %>% 
  distinct()

# mean_tlp_angio <- field_segment_angio_df %>% 
#   drop_na(tlp) %>% 
#  # filter(year == 2021, timing == "fall") %>% 
#   mutate(mean_tlp_all = mean(tlp)) %>% 
#   select(mean_tlp_all) %>% 
#   distinct() %>% 
#   pull()
# mean_tlp_angio

# sd_tlp_angio <- field_segment_angio_df %>% 
#   drop_na(tlp) %>% 
#  # filter(year == 2021, timing == "fall") %>% 
#   mutate(sd_tlp_all = sd(tlp)) %>% 
#   select(sd_tlp_all) %>% 
#   distinct() %>% 
#   pull() 
# sd_tlp_angio

# upr_tlp_angio <- mean(mean_tlp_angio + (sd_tlp_angio/2))
# upr_tlp_angio
# lwr_tlp_angio <- mean(mean_tlp_angio - (sd_tlp_angio/2))

angio_seg_plot <- ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_segment_angio_df) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_norand_df_angio$lwr_angio, 
                  xmax = seg_psi_norand_df_angio$upr_angio), 
              fill = "red", 
              color = "red", 
              alpha = .5) +
  geom_ribbon(aes(xmin = seg_psi_df_angio$lwr_angio, 
                  xmax = seg_psi_df_angio$upr_angio), 
              fill = "pink", 
              color = "pink", 
              alpha = .5)   +
  geom_vline(xintercept  = seg_psi_df_angio$mpa_psi_angio, color = "pink", size = 2)  +
   geom_vline(xintercept  = seg_psi_norand_df_angio$mpa_psi_angio, color = "red", size = 2) +
  geom_vline(xintercept  = seg_psi_norand_df_angio$mpa_psi_angio, linetype = "dotted") +
  #geom_vline(xintercept  = field_data_angio$tlp, colour = "blue") +
 # geom_vline(xintercept  = field_data_gymno$tlp, colour = "purple") +
  geom_vline(xintercept  = mean_tlp_angio, colour = "orange", size = 2) +
  geom_ribbon(aes(xmin = lwr_tlp_angio, 
                  xmax = upr_tlp_angio), 
              fill = "orange", 
              color = "orange", 
              alpha = .5)  +
  labs(title = "Angiosperms", 
       #caption = "pink = est. breakpoint with no random effects, red = with random effects, orange = tlp"
       x = "", 
       y = "LFM") +
  theme(legend.position = "none") 

angio_seg_plot
```

## Gymnosperms
```{r}
#Data
field_segment_gymno_df <-flam_curve_df_4%>% 
  filter(F.Group == "Gymnosperm") %>% 
  mutate(id = plant,
         mpa = water_potential) %>% 
  drop_na(mpa, lfm)

mpa_psi <- c(rep(NaN, 4))
mpa_se <- c(rep(NaN, 4))
seg_psi_df <- tibble::data_frame(mpa_psi, mpa_se)


mpa_psi <- c(rep(NaN, 4))
mpa_se <- c(rep(NaN, 4))
mpa_p <- c(rep(NaN, 4))
seg_psi_norand_df <- tibble::data_frame(mpa_psi, mpa_se, mpa_p)

#With random
field_segment_gymno_df <- field_segment_gymno_df %>% 
  drop_na(mpa, lfm) %>% 
  mutate(id = plant)

o_mod<-lme(lfm ~ mpa + spp, random=~1|id, data=field_segment_gymno_df)

os_mod<-segmented.default(o_mod, ~mpa, npsi=list(lfm=1))
#summarizing results (note the '_coef' argument)
summary.segmented(os_mod)

#put into the df:
seg_psi_df[1] <- summary(os_mod)$psi[2]
seg_psi_df[2] <- summary(os_mod)$psi[3]

#No random
u_mod <- lm(lfm~mpa+ spp, data = field_segment_gymno_df)
u_mod_b <- segmented(u_mod, ~mpa)
summary.segmented(u_mod_b)

slopes_gymno <- slope(u_mod_b)

u_mod_davies <- davies.test(u_mod, seg.Z = ~mpa)
u_mod_davies

u_mod_pscore <- pscore.test(u_mod_b)
u_mod_pscore

#put into the df:
seg_psi_norand_df[1] <- summary(u_mod_b)$psi[2]
seg_psi_norand_df[2] <- summary(u_mod_b)$psi[3]
seg_psi_norand_df[3] <- u_mod_davies$p.value


#Set up SE grey areas: 

seg_psi_df <- seg_psi_df %>% 
  mutate(lwr = mpa_psi - (mpa_se/2), 
         upr = mpa_psi + (mpa_se/2)) %>% 
  distinct()

seg_psi_norand_df <- seg_psi_norand_df %>% 
  mutate(lwr = mpa_psi - (mpa_se/2), 
         upr = mpa_psi + (mpa_se/2)) %>% 
  distinct()

# mean_tlp <- field_segment_gymno_df %>% 
#   drop_na(tlp) %>% 
#  # filter(year == 2021, timing == "fall") %>% 
#   mutate(mean_tlp_all = mean(tlp)) %>% 
#   select(mean_tlp_all) %>% 
#   distinct() %>% 
#   pull()
# mean_tlp
# 
# sd_tlp <- field_segment_gymno_df %>% 
#   drop_na(tlp) %>% 
#  # filter(year == 2021, timing == "fall") %>% 
#   mutate(sd_tlp_all = sd(tlp)) %>% 
#   select(sd_tlp_all) %>% 
#   distinct() %>% 
#   pull() 
# sd_tlp
# 
# upr_tlp <- mean(mean_tlp + (sd_tlp/2))
# upr_tlp
# lwr_tlp <- mean(mean_tlp - (sd_tlp/2))

gymno_seg_plot <- ggplot(aes(x = mpa, 
           y = lfm, 
           color = spp), 
           data = field_segment_gymno_df) +
  geom_point()+
  geom_ribbon(aes(xmin = seg_psi_norand_df$lwr, 
                  xmax = seg_psi_norand_df$upr), 
              fill = "red", 
              color = "red", 
              alpha = .5) +
 geom_ribbon(aes(xmin = seg_psi_df$lwr,
                 xmax = seg_psi_df$upr),
             fill = "pink",
             color = "pink",
             alpha = .5)   +
 geom_vline(xintercept  = seg_psi_df$mpa_psi, color = "pink", size = 2) +
   geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, color = "red", size = 2) +
  geom_vline(xintercept  = seg_psi_norand_df$mpa_psi, linetype = "dotted") +
  #geom_vline(xintercept  = field_data_gymno$tlp, colour = "blue") +
 # geom_vline(xintercept  = field_data_gymno$tlp, colour = "purple") +
  geom_vline(xintercept  = mean_tlp, colour = "orange", size = 2) +
  geom_ribbon(aes(xmin = lwr_tlp, 
                  xmax = upr_tlp), 
              fill = "orange", 
              color = "orange", 
              alpha = .5)  +
  labs(title = "Gymnosperms", 
       caption = "pink = est. breakpoint with no random effects, red = with random effects, orange = tlp") +
  theme(legend.position = "none") 

gymno_seg_plot

```
